Fix resultado_DB

+++ Code/AgroSolum/core/models.py
@@ -2,8 +2,8 @@ from django.db import models
 
 class Produtos(models.Model):
     descricao = models.CharField( max_length=100)  
-    quantidade = models.IntegerField()
-    valor = models.IntegerField()
+    quantidade = models.FloatField()
+    valor = models.FloatField()
     vencimento = models.DateField()
     
 class API(models.Model):
